[gaim-migrate @ 8088] I made Gaim compile with -ansi in my CFLAGS... I don't really know why. I had to set HAVE_GETADDRINFO and HAVE_SIGNAL_H to 0
--HG--
extra : convert_revision : bfeed7e8ab2319644d3e2630eec484bb2784e45d

+++ src/protocols/zephyr/zephyr.c
@@ -366,7 +366,7 @@ static void handle_message(ZNotice_t notice, struct sockaddr_in from)
 		char *realmptr;
 		char *sendertmp;
 		char *ptr = notice.z_message + strlen(notice.z_message) + 1;
-		int len = notice.z_message_len - (ptr - notice.z_message);
+		int len = notice.z_message_len - ((int)ptr - (int)notice.z_message);
 		GaimConvImFlags flags = 0;
 		if (len > 0) {
 			buf = g_malloc(len + 1);
